Refactors the navigation drawer to remain consistent with the filter drawer refactor
